Refactor into multiple files and promptly break everything
